backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.12.5
pydantic-settings==2.5.2
pyjwt==2.10.1
supabase==2.25.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.9

